Move file tree deletion tests into file-tree.spect.ts
